fixed #2454 (https://beyond-aion.com/forum/thread/2454-nach-ruhnatorium-gruppe-nicht-aufzulösen/) autogroups not disbanding after instance finish && minor refactorings

+++ game-server/src/com/aionemu/gameserver/network/aion/clientpackets/CM_AUTO_GROUP.java
@@ -56,7 +56,7 @@ public class CM_AUTO_GROUP extends AionClientPacket {
 				AutoGroupService.getInstance().cancelEnter(player, instanceMaskId);
 				break;
 			case 104:
-				PeriodicInstanceManager.getInstance().handleRequest(player, (byte) instanceMaskId);
+				PeriodicInstanceManager.getInstance().handleRequest(player, instanceMaskId);
 				break;
 			case 105:
 				// DredgionRegService.getInstance().failedEnterDredgion(player);
